+++ src/cli.h
@@ -11,6 +11,9 @@
 SEXP clic_diff_chr(SEXP a, SEXP b, SEXP max);
 
 SEXP clic_md5(SEXP strs);
+SEXP clic_md5_raw(SEXP r);
+SEXP clic_sha256(SEXP strs);
+SEXP clic_sha256_raw(SEXP r);
 SEXP clic_tty_size();
 SEXP clic_ansi_simplify(SEXP x, SEXP keep_csi);
 SEXP clic_ansi_substr(SEXP x, SEXP start, SEXP stop);